requirements.txt
auto_mix_prep==0.2.0
fastapi==0.115.6
funasr==1.1.16
Jinja2==3.1.4
moviepy==1.0.3
openai==1.57.0
Pillow==11.0.0
pydantic==2.10.3
pydub==0.25.1
qrcode==8.0
Requests==2.32.3
starlette==0.41.3
supabase==2.10.0
torch==2.5.1
uvicorn==0.32.1
